Added Ability to Ban users
Now with 100% more banhammer usage!

Ability to ban or unban a user with a click from the user's list screen.
Only thing that would make it easier was if I thickboxed it so that it
didn't take you away from the screen.

The ban message is configurable using the SEF_BAN_MESSAGE define.